Source code dump
This is a simple version of libocn that doesn't really do anything.
The general idea is to flesh this out a bit with an all pairs shortest
path algorithm so that larger mesh networks will have their full paths
traversable, at which point I should be able to use this for
Jonathan's DREAMER toolchain.