Sort cmesh so it maps like a mesh
The general idea here is that I want to map networks that aren't just
meshes to the placement tool used by the mesh network.  The other
network types that are supported are easy: mesh just maps directly and
it doesn't matter for a crossbar.  This makes cmesh map pretty well.